// src/app/interfaces/common/project.interface.ts
export interface Project {
  _id?: string;
  name?: string;
  domain?: string;
  url?: string;
  dateString?: string;
  googleIndex?: string;
  semRushBackLinks?: string;
  semRushSubDomainBacklinks?: string;
  bingIndex?: string;
  semRushRank?: string;
  semRushSeTraffic?: string;
  projectId?: string;
  projectName?: string;
  guestPostPrice?: string;
  linkInsertPrice?: string;
  semRushUsTraffic?: string;
  trend?: string;
  mode?: string;
  ip?: string;
  ur?: string;
  ahrefsRank?: string;
  refDomainsDoFollow?: string;
  refDomainsGovernmental?: string;
  refDomainsEducational?: string;
  refIPs?: string;
  refSubNets?: string;
  linkedDomains?: string;
  totalBacklinks?: string;
  backlinksText?: string;
  backlinksNoFollow?: string;
  backlinksRedirect?: string;
  backlinksImage?: string;
  backlinksFrame?: string;
  backlinksForm?: string;
  backlinksGovernmental?: string;
  backlinksEducational?: string;
  keywordsRankedFor?: string;
  ahrefTraffic?: string;
  trafficDate?: string;
  date?: any;
  contact?: string;
  da?: string;
  niche?: string;
  secondaryNiche?: string;
  price?: string;
  linkInsert?: string;
  condition?: string;
  replyStatus?: string;
  dr?: string;
  traffic?: string;
  keywords?: string;
  status?: string;
  select?: boolean;
  createdAt?: Date;
  updatedAt?: Date;
}